Fixed Low Dilution Issue
Fixed the problem so that instead of increasing the top up volume by a large amount, instead I changed it to force the factor added become a set amount and increasing the media used. This method uses much less media and factor and is also the logical way to move forward.